Split diff and patch into separate files.
